refactor(macos): finish terminal pane view split
