+++ clients/apple/ARCHITECTURE.md
@@ -32,12 +32,14 @@ Xcode target.
 | `Views/Shell/ShellPaneTopologyIndicator.swift` | 285 | SwiftUI; macOS gates | Pane topology rendering and direct split-pane focus actions | `Views/Shell/` |
 | `Views/Shell/ShellWorkspaceView.swift` | 46 | SwiftUI; macOS gates | Shell workspace composition and space keyboard shortcuts | `Views/Shell/` |
 | `Views/Shell/ShellCommandTabView.swift` | 621 | SwiftUI; macOS gates | Command palette search, routing, attention, and action presentation | `Views/Shell/` |
-| `TerminalPaneView.swift` | 1337 | Foundation, SwiftUI; macOS gates | Workspace panel composition, empty-state presentation, pane title bars, find, and local diagnostic helpers | `Views/Shell/Terminal/` |
+| `TerminalPaneView.swift` | 229 | Foundation, SwiftUI; macOS gates | Workspace canvas composition, preview selection, and empty-state panel presentation | `Views/Shell/Terminal/` |
 | `Views/Shell/Settings/ShellSettingsContentView.swift` | 496 | Foundation, OSLog, SwiftUI; macOS gates | Settings composition, state refresh, managed-user actions, and diagnostics export | `Views/Shell/Settings/` |
 | `Views/Shell/Settings/ShellSettingsNavigationView.swift` | 190 | SwiftUI; macOS gates | Settings navigation rail, group selection, and pane backgrounds | `Views/Shell/Settings/` |
 | `Views/Shell/Settings/ShellSettingsComponents.swift` | 644 | SwiftUI; macOS gates | Settings rows, action accessories, managed-user sheets, metrics, and typography | `Views/Shell/Settings/` |
 | `Views/Shell/Terminal/ShellPaneTreeLayoutView.swift` | 363 | Foundation, SwiftUI; macOS gates | Pane-tree dispatch, split layout, resize gestures, and leaf action wiring | `Views/Shell/Terminal/` |
-| `Views/Shell/Terminal/ShellTerminalLeafView.swift` | 173 | Foundation, SwiftUI; macOS gates | Terminal leaf composition, restored transcript presentation, and pane-scoped find overlay wiring | `Views/Shell/Terminal/` |
+| `Views/Shell/Terminal/ShellTerminalLeafView.swift` | 171 | Foundation, SwiftUI; macOS gates | Terminal leaf composition, restored transcript presentation, and pane-scoped find overlay wiring | `Views/Shell/Terminal/` |
+| `Views/Shell/Terminal/ShellPaneTitleBarViews.swift` | 585 | Foundation, SwiftUI; macOS gates | Terminal and bounded-content title bars, responsive accessories, and activity freshness | `Views/Shell/Terminal/` |
+| `Views/Shell/Terminal/ShellTerminalOverlayViews.swift` | 128 | SwiftUI; macOS gates | Pane-scoped Find interaction and passive inactive-pane dimming | `Views/Shell/Terminal/` |
 | `Views/Shell/Content/ShellBoundedContentViews.swift` | 647 | Foundation, SwiftUI; macOS gates | Bounded Agent and Markdown content renderers plus unavailable-content presentation | `Views/Shell/Content/` |
 | `TerminalHostView.swift` | 1911 | AppKit, Carbon, QuartzCore, GhosttyKit; macOS gates | AppKit terminal host bridge, focus, overlay composition, runtime attachment, and collaborator wiring | `Views/Shell/Terminal/` plus terminal collaborators |
 | `GhosttyLiveHost.swift` | 1227 | Foundation, AppKit, GhosttyKit, OSLog, QuartzCore; macOS/Ghostty gates | Ghostty canvas bridge and wakeup/occlusion integration | `Services/Terminal/` or `Support/TerminalBridge/` |
@@ -319,12 +321,13 @@ device support was not required for this validation.
 ## Remaining Architecture Debt
 
 `check-architecture-maintainability.sh` currently completes in report mode.
-7 known large-file / bridge-boundary warnings remain after the first ten Apple
+6 known large-file / bridge-boundary warnings remain after the first eleven Apple
 ownership slices removed the `ShellHostController.swift` bridge warning and
 split control projection, observation commands, root-controller
 responsibilities, shell presentation models, settings models, snapshot DTO
 families, shell/platform value families, sidebar and settings presentation,
-and pane-tree/content rendering into focused owners. These warnings
+pane-tree/content rendering, and pane title/overlay presentation into focused
+owners. These warnings
 are telemetry for the cleanup, not the cleanup
 definition. The real debt is any Swift production source that still carries a
 Rust-owned shell-domain implementation, fixture, or fallback after shell-core
@@ -415,6 +418,12 @@ into focused owners under `Views/Shell/Terminal/` and `Views/Shell/Content/`.
 The warning count remains 7, while the hard `TerminalPaneView.swift` ceiling
 falls from 2,513 to 1,337 lines ahead of the title-bar, find-bar, and local
 utility split.
+
+The eleventh ownership slice moved responsive pane title bars and terminal
+overlay presentation into focused owners under `Views/Shell/Terminal/`, and
+deleted the unreferenced runtime, boot, Ghostty, binding, action, chip, and info
+card presentation chain. `TerminalPaneView.swift` falls from 1,337 to 229 lines
+and exits the large-file ledger, lowering the warning count from 7 to 6.
 
 Rust-owned Swift legacy cleanup targets at this baseline:
