+++ clients/apple/README.md
@@ -24,7 +24,8 @@ recorded in [`ARCHITECTURE.md`](./ARCHITECTURE.md).
 - `MacShellRootView.swift`: thin primary macOS shell composition root
 - `Views/Shell/`: shell composition plus focused sidebar, Space slider, tab-row,
   drop-target, activity, pane-topology, workspace, command-palette, settings,
-  pane-tree, terminal-leaf, and bounded-content SwiftUI components
+  pane-tree, terminal-leaf, pane-title, terminal-overlay, and bounded-content
+  SwiftUI components
 - `Models/Shell/`: focused shell, Terminal Profile, managed-account, activity,
   context, and pane/content/tree/tab/Space/workspace snapshot families; runtime support,
   focused presentation models, and settings navigation, Terminal Profile, managed-user,
@@ -38,8 +39,8 @@ recorded in [`ARCHITECTURE.md`](./ARCHITECTURE.md).
 - `Services/Terminal/`: terminal host runtime reporting, window observation,
   agent-activity projection, and terminal runtime service collaborators
 - `TerminalPaneView.swift` / `TerminalHostView.swift`: current terminal pane and host surfaces;
-  settings, bounded content, pane-tree, and terminal-leaf presentation live under
-  `Views/Shell/`
+  settings, bounded content, pane-tree, terminal-leaf, title-bar, and overlay
+  presentation live under `Views/Shell/`
 - `ShellHostController.swift`: observable shell state and root controller lifecycle
 - `ShellControlPlane.swift`: thin shell control-plane orchestration across the
   shell service owners
